modularize tokenizer func & class Stack
